fastapi_server/requirements.txt
fastapi
uvicorn
pandas
requests
haversine
python-dotenv